cmd: Add round-trip test
Add some test cases for data being round tripped correctly, to increase
confidence when making changes and not needing to rely on running it
locally against a real Kubernetes cluster.

`EDITOR` (which takes multiple args) is slightly abused to modify the
file without needing to hijack STDIN/STDOUT. The original environment
variable state is restored at the end of the test, so as not to affect
other tests/commands.

Logging is discarded because it makes the output harder to read for
passing tests and I couldn't find a way to only capture it for failing
tests.