cmd: Improve handling of empty secrets
Previously if you created an empty secret and then used `modify-secret`
to add data to it, you were presented with an empty JSON document, which
was confusing when you were expecting to write YAML:

    $ kubectl create secret generic dcarley
    secret/dcarley created
    $ EDITOR=cat kubectl modify-secret dcarley
    {}
    INFO[0000] no changes done to secret "dcarley"

This seems to be a "feature" of the YAML library when you marshal an
empty map:

- https://play.golang.org/p/9Byzr0yfQD3

Prevent this from happening by passing an empty byte slice (and thus
file) if the map is empty. The existing checksum continues to work since
we only checksum after writing `yamlData`.

I've pulled this out to a function so that it's easier to document and
test in isolation. It wasn't possible to test this with the round-trip
because we don't have a way to inspect the file in the middle of the
operation.